requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
anthropic==0.29.0
google-generativeai==0.4.1